+++ src/nectarapi/graphenerpc.py
@@ -2,7 +2,9 @@
 import json
 import logging
 import re
-import ssl
+
+import requests
+from requests.exceptions import ConnectionError
 
 from nectargraphenebase.chains import known_chains
 from nectargraphenebase.version import version as nectar_version
@@ -17,30 +19,6 @@ from .exceptions import (
 )
 from .node import Nodes
 from .rpcutils import get_api_name, get_query, is_network_appbase_ready
-
-WEBSOCKET_MODULE = None
-if not WEBSOCKET_MODULE:
-    try:
-        import websocket
-        from websocket._exceptions import (
-            WebSocketConnectionClosedException,
-            WebSocketTimeoutException,
-        )
-
-        WEBSOCKET_MODULE = "websocket"
-    except ImportError:
-        WEBSOCKET_MODULE = None
-REQUEST_MODULE = None
-if not REQUEST_MODULE:
-    try:
-        import requests
-        from requests.adapters import HTTPAdapter
-        from requests.exceptions import ConnectionError
-        from requests.packages.urllib3.util.retry import Retry
-
-        REQUEST_MODULE = "requests"
-    except ImportError:
-        REQUEST_MODULE = None
 
 log = logging.getLogger(__name__)
 
@@ -58,23 +36,9 @@ def set_session_instance(instance):
 
 def shared_session_instance():
     """Get session instance"""
-    if REQUEST_MODULE is None:
-        raise Exception("Requests module is not available.")
     if not SessionInstance.instance:
         SessionInstance.instance = requests.Session()
     return SessionInstance.instance
-
-
-def create_ws_instance(use_ssl=True, enable_multithread=True):
-    """Get websocket instance"""
-    if WEBSOCKET_MODULE is None:
-        raise Exception("WebSocket module is not available.")
-    if use_ssl:
-        ssl_defaults = ssl.get_default_verify_paths()
-        sslopt_ca_certs = {"ca_certs": ssl_defaults.cafile}
-        return websocket.WebSocket(sslopt=sslopt_ca_certs, enable_multithread=enable_multithread)
-    else:
-        return websocket.WebSocket(enable_multithread=enable_multithread)
 
 
 class GrapheneRPC(object):
@@ -83,7 +47,7 @@ class GrapheneRPC(object):
 
     It logs warnings and errors.
 
-    :param str urls: Either a single Websocket/Http URL, or a list of URLs
+    :param str urls: Either a single HTTP URL, or a list of HTTP URLs
     :param str user: Username for Authentication
     :param str password: Password for Authentication
     :param int num_retries: Number of retries for node connection (default is 100)
@@ -97,8 +61,8 @@ class GrapheneRPC(object):
 
     def __init__(self, urls, user=None, password=None, **kwargs):
         """Initialize the RPC client."""
-        self.rpc_methods = {"offline": -1, "ws": 0, "jsonrpc": 1, "wsappbase": 2, "appbase": 3}
-        self.current_rpc = self.rpc_methods["ws"]
+        self.rpc_methods = {"offline": -1, "appbase": 3}
+        self.current_rpc = self.rpc_methods["appbase"]
         self._request_id = 0
         self.timeout = kwargs.get("timeout", 60)
         num_retries = kwargs.get("num_retries", 100)
@@ -119,7 +83,6 @@ class GrapheneRPC(object):
 
         self.user = user
         self.password = password
-        self.ws = None
         self.url = None
         self.session = None
         self.rpc_queue = []
@@ -149,16 +112,11 @@ class GrapheneRPC(object):
 
     def next(self):
         """Switches to the next node url"""
-        if self.ws:
-            try:
-                self.rpcclose()
-            except Exception as e:
-                log.warning(str(e))
         self.rpcconnect()
 
     def is_appbase_ready(self):
         """Check if node is appbase ready"""
-        return self.current_rpc in [self.rpc_methods["wsappbase"], self.rpc_methods["appbase"]]
+        return self.current_rpc == self.rpc_methods["appbase"]
 
     def get_use_appbase(self):
         """Returns True if appbase ready and appbase calls are set"""
@@ -173,36 +131,21 @@ class GrapheneRPC(object):
                 self.url = next(self.nodes)
                 self.nodes.reset_error_cnt_call()
                 log.debug("Trying to connect to node %s" % self.url)
-                if self.url[:3] == "wss":
-                    self.ws = create_ws_instance(use_ssl=True)
-                    self.ws.settimeout(self.timeout)
-                    self.current_rpc = self.rpc_methods["wsappbase"]
-                elif self.url[:2] == "ws":
-                    self.ws = create_ws_instance(use_ssl=False)
-                    self.ws.settimeout(self.timeout)
-                    self.current_rpc = self.rpc_methods["wsappbase"]
-                else:
-                    self.ws = None
-                    self.session = shared_session_instance()
-                    if self.use_tor:
-                        self.session.proxies = {}
-                        self.session.proxies["http"] = "socks5h://localhost:9050"
-                        self.session.proxies["https"] = "socks5h://localhost:9050"
-                    self.current_rpc = self.rpc_methods["appbase"]
-                    self.headers = {
-                        "User-Agent": "nectar v%s" % (nectar_version),
-                        "content-type": "application/json; charset=utf-8",
-                    }
+                self.ws = None
+                self.session = shared_session_instance()
+                if self.use_tor:
+                    self.session.proxies = {}
+                    self.session.proxies["http"] = "socks5h://localhost:9050"
+                    self.session.proxies["https"] = "socks5h://localhost:9050"
+                self.current_rpc = self.rpc_methods["appbase"]
+                self.headers = {
+                    "User-Agent": "nectar v%s" % (nectar_version),
+                    "content-type": "application/json; charset=utf-8",
+                }
             try:
-                if self.ws:
-                    self.ws.connect(self.url)
-                    self.rpclogin(self.user, self.password)
                 if self.disable_chain_detection:
                     # Set to appbase rpc format
-                    if self.current_rpc == self.rpc_methods["ws"]:
-                        self.current_rpc = self.rpc_methods["wsappbase"]
-                    else:
-                        self.current_rpc = self.rpc_methods["appbase"]
+                    self.current_rpc = self.rpc_methods["appbase"]
                     break
                 try:
                     props = None
@@ -213,18 +156,12 @@ class GrapheneRPC(object):
                 except Exception as e:
                     if re.search("Bad Cast:Invalid cast from type", str(e)):
                         # retry with not appbase
-                        if self.current_rpc == self.rpc_methods["wsappbase"]:
-                            self.current_rpc = self.rpc_methods["ws"]
-                        else:
-                            self.current_rpc = self.rpc_methods["appbase"]
+                        self.current_rpc = self.rpc_methods["appbase"]
                         props = self.get_config(api="database")
                 if props is None:
                     raise RPCError("Could not receive answer for get_config")
                 if is_network_appbase_ready(props):
-                    if self.ws:
-                        self.current_rpc = self.rpc_methods["wsappbase"]
-                    else:
-                        self.current_rpc = self.rpc_methods["appbase"]
+                    self.current_rpc = self.rpc_methods["appbase"]
                 break
             except KeyboardInterrupt:
                 raise
@@ -233,18 +170,6 @@ class GrapheneRPC(object):
                 do_sleep = not next_url or (next_url and self.nodes.working_nodes_count == 1)
                 self.nodes.sleep_and_check_retries(str(e), sleep=do_sleep)
                 next_url = True
-
-    def rpclogin(self, user, password):
-        """Login into Websocket"""
-        if self.ws and self.current_rpc == self.rpc_methods["ws"] and user and password:
-            self.login(user, password, api="login_api")
-
-    def rpcclose(self):
-        """Close Websocket"""
-        if self.ws is None:
-            return
-        # if self.ws.connected:
-        self.ws.close()
 
     def request_send(self, payload):
         if self.user is not None and self.password is not None:
@@ -262,13 +187,6 @@ class GrapheneRPC(object):
         if response.status_code == 401:
             raise UnauthorizedError
         return response
-
-    def ws_send(self, payload):
-        if self.ws is None:
-            raise RPCConnection("No websocket available!")
-        self.ws.send(payload)
-        reply = self.ws.recv()
-        return reply
 
     def version_string_to_int(self, network_version):
         version_list = network_version.split(".")
@@ -445,16 +363,10 @@ class GrapheneRPC(object):
         while True:
             self.nodes.increase_error_cnt_call()
             try:
-                if (
-                    self.current_rpc == self.rpc_methods["ws"]
-                    or self.current_rpc == self.rpc_methods["wsappbase"]
-                ):
-                    reply = self.ws_send(json.dumps(payload, ensure_ascii=False).encode("utf8"))
-                else:
-                    response = self.request_send(
-                        json.dumps(payload, ensure_ascii=False).encode("utf8")
-                    )
-                    reply = response.text
+                response = self.request_send(
+                    json.dumps(payload, ensure_ascii=False).encode("utf8")
+                )
+                reply = response.text
                 if not bool(reply):
                     try:
                         self.nodes.sleep_and_check_retries("Empty Reply", call_retry=True)
@@ -468,15 +380,7 @@ class GrapheneRPC(object):
                     break
             except KeyboardInterrupt:
                 raise
-            except WebSocketConnectionClosedException as e:
-                self.nodes.increase_error_cnt()
-                self.nodes.sleep_and_check_retries(str(e), sleep=False, call_retry=False)
-                self.rpcconnect()
             except ConnectionError as e:
-                self.nodes.increase_error_cnt()
-                self.nodes.sleep_and_check_retries(str(e), sleep=False, call_retry=False)
-                self.rpcconnect()
-            except WebSocketTimeoutException as e:
                 self.nodes.increase_error_cnt()
                 self.nodes.sleep_and_check_retries(str(e), sleep=False, call_retry=False)
                 self.rpcconnect()